create seperate view component service
