+++ biosimulations/apps/dispatch/src/app/components/simulations/view/view.component.ts
@@ -36,7 +36,7 @@ import { ConfigService } from '@biosimulations/shared/services';
 import { BehaviorSubject, interval, Observable, of, Subject, Subscription } from 'rxjs';
 import { concatAll, flatMap, map, repeat, shareReplay, takeUntil, takeWhile, tap } from 'rxjs/operators';
 import { AxisLabelType, AXIS_LABEL_TYPES, CombineArchive, DataSetIdDisabled, FormattedSimulation, Report, ScatterTraceModeLabel, SCATTER_TRACE_MODEL_LABELS } from './view.model'
-
+import { ViewService } from './view.service'
 @Component({
   templateUrl: './view.component.html',
   styleUrls: ['./view.component.scss'],
@@ -85,6 +85,7 @@ export class ViewComponent implements OnInit, OnDestroy {
   constructor(
     private config: ConfigService,
     private route: ActivatedRoute,
+    private service: ViewService,
     private formBuilder: FormBuilder,
     private simulationService: SimulationService,
     private appService: VisualizationService,
@@ -111,7 +112,7 @@ export class ViewComponent implements OnInit, OnDestroy {
     this.formGroup.controls.yDataSetIds.disable();
 
     this.Simulation$ = this.simulationService.getSimulation(this.uuid).pipe(shareReplay(1))
-    this.formattedSimulation$ = this.Simulation$.pipe(map<Simulation, FormattedSimulation>(this.formatSimulation.bind(this)))
+    this.formattedSimulation$ = this.Simulation$.pipe(map<Simulation, FormattedSimulation>(this.service.formatSimulation))
     this.statusRunning$ = this.formattedSimulation$.pipe(map(value => SimulationStatusService.isSimulationStatusRunning(value.status)))
     this.statusSuceeded$ = this.formattedSimulation$.pipe(map(value => SimulationStatusService.isSimulationStatusSucceeded(value.status)))
     this.logs$ = this.statusRunning$.pipe(map(running => {
@@ -136,44 +137,6 @@ export class ViewComponent implements OnInit, OnDestroy {
     this.subs.forEach(subscription => subscription.unsubscribe())
   }
 
-  private formatSimulation(simulation: Simulation): FormattedSimulation {
-
-    const statusRunning = SimulationStatusService.isSimulationStatusRunning(
-      simulation.status
-    );
-    const statusSucceeded = SimulationStatusService.isSimulationStatusSucceeded(
-      simulation.status);
-    return {
-      id: simulation.id,
-      name: simulation.name,
-      simulator: simulation.simulator,
-      simulatorVersion: simulation.simulatorVersion,
-      status: simulation.status,
-      statusRunning: statusRunning,
-      statusSucceeded: statusSucceeded,
-      statusLabel: SimulationStatusService.getSimulationStatusMessage(
-        simulation.status,
-        true
-      ),
-      runtime:
-        simulation.runtime !== undefined
-          ? Math.round(simulation.runtime / 1000).toString() + ' s'
-          : 'N/A',
-      submitted: new Date(simulation.submitted).toLocaleString(),
-      updated: new Date(simulation.updated).toLocaleString(),
-      projectSize:
-        simulation.projectSize !== undefined
-          ? (simulation.projectSize / 1024).toFixed(2) + ' KB'
-          : '',
-      resultsSize:
-        simulation.resultsSize !== undefined
-          ? (simulation.resultsSize / 1024).toFixed(2) + ' KB'
-          : 'N/A',
-      projectUrl: `${urls.dispatchApi}run/${simulation.id}/download`,
-      simulatorUrl: `${this.config.simulatorsAppUrl}simulators/${simulation.simulator}/${simulation.simulatorVersion}`,
-      resultsUrl: `${urls.dispatchApi}download/result/${simulation.id}`
-    }
-  }
 
 
   private setProjectOutputs(combineArchive: CombineArchive): void {
