+++ biosimulations/apps/dispatch/src/app/components/simulations/view/view.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ViewService } from './view.service';
+
+describe('ViewService', () => {
+  let service: ViewService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ViewService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
